Day 44: Processor Code Restructuring
Each processor lives in its own cpp/h pair and is appropriately
included and built, reducing 'where is the x' confusion that
SC originally had at the cost of a bushier filesystem.
