+++ src/CMakeLists.txt
@@ -7,9 +7,10 @@ add_library(${PROJECT_NAME} STATIC
         dsp/data_tables.cpp
         dsp/processor/processor.cpp
 
-        dsp/processor/processor_supersvf.cpp
-        dsp/processor/processor_oscillators.cpp
-        dsp/processor/processor_distortions.cpp
+        dsp/processor/distortion/microgate.cpp
+        dsp/processor/filter/supersvf.cpp
+        dsp/processor/oscillator/oscpulsesync.cpp
+        dsp/processor/oscillator/phasemodulation.cpp
 
         engine/engine.cpp
         engine/zone.cpp
